Split main file into multiple files
- Update tests to remove references to nested classes and
  unused imports
- Take care of some deviations from standards to avoid circular
  imports:
  - In connectors.py: Media is imported in Connect.create()
  - In pages.py: file_to_xml() is imported in Page.set_name()

  It might be best to try to break these circular imports in a
  cleaner way, by re-working the code or user's workflow/interface.
